+++ classes/Event/Storing.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Neat\Object\Event;
+
+use Neat\Object\Event;
+
+class Storing extends Event
+{
+}