+++ changelog.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Support data to be merged in sql query when using Repository::sql().
+- PSR-14 Event Dispatcher support using repository decorator.
+- Event implementations: Loading, Loaded, Storing, Stored, Deleting, Deleted.
+
+### Removed
+- PHP 7.0 and PHP 7.1 support.
 
 ## [0.11.0] - 2020-06-19
 ### Changed